fix incorrect use of "inject"
"inject", as opposed to "spawn" has a clearly defined meaning per
doc/TCL.md. The usage of the term "inject" was not consistently used,
and in several placed was used to refer to spawning. This has been
rectified.

Namely, this renames the existing "inject" instrument to "spawn". A new
inject instrument is added to align with the proper meaning. Also, the
"inject" command is now named "spawn", although an alias named "inject"
is preserved for convenience.

Additionally, many performance counters are added both globally and to
notes.

+++ src/vec.h
@@ -109,10 +109,10 @@
 
 #define vec_find(v, val, idx)\
 		do {\
-			for ((idx) = 0; (idx) < (v)->length; (idx)++) {\
+			for ((idx) = 0; ((unsigned int) idx) < (v)->length; (idx)++) {\
 				if ((v)->data[(idx)] == (val)) break;\
 			}\
-			if ((idx) == (v)->length) (idx) = -1;\
+			if (((unsigned int) idx) == (v)->length) (idx) = -1;\
 		} while (0)
 
 
@@ -126,7 +126,7 @@
 
 #define vec_reverse(v)\
 		do {\
-			int i__ = (v)->length / 2;\
+			unsigned int i__ = (v)->length / 2;\
 			while (i__--) {\
 				vec_swap((v), i__, (v)->length - (i__ + 1));\
 			}\
